+++ packages/kit/types/private.d.ts
@@ -26,6 +26,14 @@ export interface AdapterEntry {
 	}) => MaybePromise<void>;
 }
 
+export type BodyValidator<T> = {
+	[P in keyof T]: T[P] extends { [k: string]: infer V }
+		? BodyValidator<V> // recurse when T[P] is an object
+		: T[P] extends BigInt | Function | Symbol
+		? never
+		: T[P];
+};
+
 // Based on https://github.com/josh-hemphill/csp-typed-directives/blob/latest/src/csp.types.ts
 //
 // MIT License
@@ -198,24 +206,9 @@ export interface PrerenderErrorHandler {
 
 export type PrerenderOnErrorValue = 'fail' | 'continue' | PrerenderErrorHandler;
 
-export interface RequestEvent<Params extends Record<string, string> = Record<string, string>> {
-	clientAddress: string;
-	locals: App.Locals;
-	params: Params;
-	platform: Readonly<App.Platform>;
-	request: Request;
-	routeId: string | null;
-	url: URL;
-}
-
 export interface RequestOptions {
 	getClientAddress: () => string;
 	platform?: App.Platform;
-}
-
-export interface ResolveOptions {
-	ssr?: boolean;
-	transformPage?: ({ html }: { html: string }) => MaybePromise<string>;
 }
 
 /** `string[]` is only for set-cookie, everything else must be type of `string` */